replaced edt import mess with get_edt
