+++ src/porespy/filters/_snows.py
@@ -18,12 +18,8 @@ from porespy.tools import (
     get_tqdm,
     ps_rect,
     ps_round,
+    get_edt,
 )
-
-try:
-    from pyedt import edt
-except ModuleNotFoundError:
-    from edt import edt
 
 
 __all__ = [
@@ -38,6 +34,7 @@ __all__ = [
 ]
 
 
+edt = get_edt()
 tqdm = get_tqdm()
 logger = logging.getLogger(__name__)
