+++ src/porespy/generators/_pseudo_packings.py
@@ -14,12 +14,8 @@ from porespy.tools import (
     get_tqdm,
     ps_round,
     unpad,
+    get_edt,
 )
-
-try:
-    from pyedt import edt
-except ModuleNotFoundError:
-    from edt import edt
 
 
 __all__ = [
@@ -29,6 +25,7 @@ __all__ = [
 ]
 
 
+edt = get_edt()
 tqdm = get_tqdm()
 logger = logging.getLogger(__name__)
