+++ src/porespy/io/_funcs.py
@@ -1,19 +1,15 @@
 import os
 import subprocess
-
 import numpy as np
 import scipy.ndimage as nd
 import skimage.measure as ms
 from skimage.morphology import ball
-
 from porespy.filters import reduce_peaks
 from porespy.networks import generate_voxel_image
-from porespy.tools import sanitize_filename
+from porespy.tools import sanitize_filename, get_edt
 
-try:
-    from pyedt import edt
-except ModuleNotFoundError:
-    from edt import edt
+
+edt = get_edt()
 
 
 def dict_to_vtk(data, filename, voxel_size=1, origin=(0, 0, 0)):
