+++ src/porespy/metrics/_funcs.py
@@ -20,11 +20,8 @@ from porespy.tools import (
     extend_slice,
     get_tqdm,
     im_to_slabs,
+    get_edt
 )
-try:
-    from pyedt import edt
-except ModuleNotFoundError:
-    from edt import edt
 
 
 __all__ = [
@@ -47,6 +44,7 @@ __all__ = [
 ]
 
 
+edt = get_edt()
 tqdm = get_tqdm()
 logger = logging.getLogger(__name__)
