+++ src/porespy/tools/_funcs.py
@@ -1,24 +1,18 @@
 import logging
-
 import numpy as np
 import scipy.ndimage as spim
-from scipy.stats import rankdata
 from numba import boolean, njit
 from skimage.morphology import ball, disk
 from skimage.segmentation import relabel_sequential
-
-from ._utils import Results
+from ._utils import Results, get_edt
 
 try:
     from skimage.measure import marching_cubes
 except ImportError:
     from skimage.measure import marching_cubes_lewiner as marching_cubes
-try:
-    from pyedt import edt
-except ModuleNotFoundError:
-    from edt import edt
 
 
+edt = get_edt()
 logger = logging.getLogger(__name__)
 
 
@@ -46,10 +40,6 @@ __all__ = [
     'overlay',
     'randomize_colors',
     'recombine',
-    'ps_ball',
-    'ps_disk',
-    'ps_rect',
-    'ps_round',
     'subdivide',
     'tilde',
     'unpad',
@@ -834,6 +824,7 @@ def extend_slice(slices, shape, pad=1):
         a.append(slice(start, stop, None))
     return tuple(a)
 
+
 @njit
 def jit_extend_slice(slices, shape, pad=1):
     shape = np.array(shape)
@@ -843,6 +834,7 @@ def jit_extend_slice(slices, shape, pad=1):
         stop = min(s.stop + pad, shape[i])
         a.append(slice(start, stop, None))
     return (a[0], a[1], a[2])
+
 
 @njit
 def pad(img):
@@ -1126,7 +1118,7 @@ def all_to_uniform(im, scale=None):
     # Alternative, might be faster
     # im2 = rankdata(im).reshape(im.shape)
     # im = (im2 - im2.min())/(im2.max() - im2.min())*(scale[1] - scale[0]) + scale[0]
-    aargsort_im = np.argsort(np.argsort(im.flatten()))  # Twice for the inverse permutation
+    aargsort_im = np.argsort(np.argsort(im.flatten()))  # 2x forinverse permutation
     linspace_im = np.linspace(scale[0], scale[1], len(aargsort_im), endpoint=True)
     uniform_flatten_im = linspace_im[aargsort_im]
     im = np.reshape(uniform_flatten_im, im.shape)
@@ -1232,131 +1224,6 @@ def mesh_region(region: bool, strel=None, voxel_size=(1.0, 1.0, 1.0)):
     result.norm = norm
     result.val = val
     return result
-
-
-def ps_disk(r, smooth=True):
-    r"""
-    Creates circular disk structuring element for morphological operations
-
-    Parameters
-    ----------
-    r : float or int
-        The desired radius of the structuring element
-    smooth : boolean
-        Indicates whether the faces of the sphere should have the little
-        nibs (``True``) or not (``False``, default)
-
-    Returns
-    -------
-    disk : ndarray
-        A 2D numpy bool array of the structring element
-
-    Examples
-    --------
-    `Click here
-    <https://porespy.org/examples/tools/reference/ps_disk.html>`_
-    to view online example.
-
-    """
-    disk = ps_round(r=r, ndim=2, smooth=smooth)
-    return disk
-
-
-def ps_ball(r, smooth=True):
-    r"""
-    Creates spherical ball structuring element for morphological operations
-
-    Parameters
-    ----------
-    r : scalar
-        The desired radius of the structuring element
-    smooth : boolean
-        Indicates whether the faces of the sphere should have the little
-        nibs (``True``) or not (``False``, default)
-
-    Returns
-    -------
-    ball : ndarray
-        A 3D numpy array of the structuring element
-
-    Examples
-    --------
-    `Click here
-    <https://porespy.org/examples/tools/reference/ps_ball.html>`_
-    to view online example.
-
-    """
-    ball = ps_round(r=r, ndim=3, smooth=smooth)
-    return ball
-
-
-def ps_round(r, ndim, smooth=True):
-    r"""
-    Creates round structuring element with the given radius and dimensionality
-
-    Parameters
-    ----------
-    r : scalar
-        The desired radius of the structuring element
-    ndim : int
-        The dimensionality of the element, either 2 or 3.
-    smooth : boolean
-        Indicates whether the faces of the sphere should have the little
-        nibs (``True``) or not (``False``, default)
-
-    Returns
-    -------
-    strel : ndarray
-        A 3D numpy array of the structuring element
-
-    Examples
-    --------
-    `Click here
-    <https://porespy.org/examples/tools/reference/ps_round.html>`_
-    to view online example.
-
-    """
-    rad = int(np.ceil(r))
-    other = np.ones([2*rad + 1 for i in range(ndim)], dtype=bool)
-    other[tuple(rad for i in range(ndim))] = False
-    if smooth:
-        ball = edt(other) < r
-    else:
-        ball = edt(other) <= r
-    return ball
-
-
-def ps_rect(w, ndim):
-    r"""
-    Creates rectilinear structuring element with the given size and
-    dimensionality
-
-    Parameters
-    ----------
-    w : scalar
-        The desired width of the structuring element
-    ndim : int
-        The dimensionality of the element, either 2 or 3.
-
-    Returns
-    -------
-    strel : D-aNrray
-        A numpy array of the structuring element
-
-    Examples
-    --------
-    `Click here
-    <https://porespy.org/examples/tools/reference/ps_rect.html>`_
-    to view online example.
-
-    """
-    if ndim == 2:
-        from skimage.morphology import square
-        strel = square(w)
-    if ndim == 3:
-        from skimage.morphology import cube
-        strel = cube(w)
-    return strel
 
 
 def overlay(im1, im2, c):
@@ -1571,7 +1438,8 @@ def extract_regions(regions, labels: list, trim=True):
         x_min, x_max = min(s[i - 1][0].start, x_min), max(s[i - 1][0].stop, x_max)
         y_min, y_max = min(s[i - 1][1].start, y_min), max(s[i - 1][1].stop, y_max)
         if regions.ndim == 3:
-            z_min, z_max = min(s[i - 1][2].start, z_min), max(s[i - 1][2].stop, z_max)
+            z_min, z_max = \
+                min(s[i - 1][2].start, z_min), max(s[i - 1][2].stop, z_max)
     if trim:
         if regions.ndim == 3:
             bbox = bbox_to_slices([x_min, y_min, z_min, x_max, y_max, z_max])
@@ -1596,6 +1464,7 @@ def _check_for_singleton_axes(im):  # pragma: no cover
         logger.warning("Input image conains a singleton axis. Reduce"
                        " dimensionality with np.squeeze(im) to avoid"
                        " unexpected behavior.")
+
 
 @njit
 def center_of_mass(im):
