+++ test/integration/test_pc_curves_with_imbibition.py
@@ -1,10 +1,9 @@
 import numpy as np
 import porespy as ps
 import matplotlib.pyplot as plt
-try:
-    from pyedt import edt
-except ModuleNotFoundError:
-    from edt import edt
+
+
+edt = ps.get_edt()
 
 
 def test_drainage(plot=False):
